refactor(backend): move team-authz predicates out of wageController
Move `isUserMemberOfTeam` and `isOwnerOfTeam` to the middleware layer
(`src/middleware/teamAuthz.ts`) and re-export them from
`teamAuthzMiddleware.ts`. Wage controller now imports them from the
middleware instead of defining them, and `PUT /setWage` uses the
`requireTeamOwner('body.teamId')` middleware instead of an inline check.

Predicates live in their own file so tests can mock them without
replacing the middleware factories — the middleware imports them by
specifier from `./teamAuthz`, which is what vi.mock targets.

Test files updated:
- claimController.test.ts, weeklyClaimController.test.ts: import
  `isUserMemberOfTeam` from `middleware/teamAuthz` and mock that module
  instead of `wageController`.
- wageController.test.ts: add `app.use` that sets `req.address`
  (previously relied on `authorizeUser`, which isn't in the wage router
  chain) and mock `prisma.team.findUnique` for the new middleware check
  on `/setWage`.
